MAINT: Move `linker::Event` -> `Linker::Event`
