+++ src/apps/scdetect/CMakeLists.txt
@@ -11,7 +11,6 @@ set(
     detector/detectorbuilder.cpp
     detector/detectorwaveformprocessor.cpp
     detector/linker/association.cpp
-    detector/linker/event.cpp
     detector/linker/strategy.cpp
     detector/linker.cpp
     detector/pot.cpp
