+++ src/apps/scdetect/test/CMakeLists.txt
@@ -29,7 +29,6 @@ set(SOURCES_integration
   ../detector/detectorbuilder.cpp
   ../detector/detectorwaveformprocessor.cpp
   ../detector/linker/association.cpp
-  ../detector/linker/event.cpp
   ../detector/linker/strategy.cpp
   ../detector/linker.cpp
   ../detector/pot.cpp
